backend/requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
python-dotenv
schedule